Turn off the 'file' store loader and adapt the tests accordingly
The TPM2 'file' store loader shouldn't be necessary, so we attempt to
turn it off and adapt the tests to add the OpenSSL provider that
offers a full 'file' store loader.

The idea is that the OpenSSL 'file' store loader should be agnostic
enough to simply rely on available decoders and adjoined keymgmt
implementations.  This includes the TPM2 decoders, when that provider
is loaded.

Signed-off-by: Richard Levitte <richard@levitte.org>