tests: support blocked sha1
Some distributions (Fedora 41) are starting to block sha1 hashes. This
prevents some test cases.

The change is documented: https://fedoraproject.org/wiki/SHA1SignaturesGuidance.
However, there is no API provided by openssl to detect if sha1 is
supported or not. As a workaraound a scripts gets added which does that
by singing and verifying some dummy data.

Disabling sha1 by default was introduced to Fedora 41 by this commit:
https://gitlab.com/redhat-crypto/fedora-crypto-policies/-/commit/035c735a8310af5e3999c327d96ad5e354837250

Removing the following 2 lines from
/etc/crypto-policies/back-ends/opensslcnf.config
  [evp_properties]
  rh-allow-sha1-signatures = no
allows to run the tests successfully also with sha1. The test log also
shows that with sha1 supported the tests are executed and without sha1
support the tests are skipped for HASH=sha1.

Signed-off-by: Adrian Freihofer <adrian.freihofer@gmail.com>