Fixed loading of parent handles in TPM TSS2 keys
We need to make sure to be able to load TPM2 TSS2 keys which
have been created with earlier versions of the tpm2-openssl
provider or with earlier versions of the tpm2-tss-engine.

The parent handle of TPM2_TSS keys created with provider version < 1.2 or engine version < 1.2.0-rc0 have been written using ASN1_INTEGER_set
ASN1_INTEGER_set takes a (signed) long (which on 32-bit systems is 32-bit and on 64-bit systems is 64bit)
As parent handles are in the range of 0x81000000 - 0x81FFFFFF the MSB on a 32-bit system is always set, and therefore is treated as negative.
This won't be the case on 64 bit systems or in case the handle had been written using BN_set_word().

The parent handle of TPM2_TSS keys create with provider verision 1.2 - 1.3 have been written using ASN1_INTEGER_set_uint64.
These values can safely be read using BN_get_word, as the values written always were of type TPM2_HANDLE (uint32_t).

Resolves: #152
See also: #74,
https://github.com/tpm2-software/tpm2-tss-engine/pull/222, https://github.com/tpm2-software/tpm2-tools/pull/3113

Signed-off-by: Thomas Kopp <thomas.kopp@securiton.ch>